+++ code/time_type_switch2.cpp
@@ -112,7 +112,7 @@ struct VisitorFor : ShapeVisitor
     #undef  FOR_EACH_MAX
 
     const shape_kind<M>& first;
-    int                  result; 
+    size_t               result;
 
 private:
     VisitorFor& operator=(const VisitorFor&); ///< No assignment operator
@@ -131,7 +131,7 @@ struct Visitor : ShapeVisitor
     #undef  FOR_EACH_MAX
 
     const Shape& second;
-    int          result;
+    size_t       result;
 
 private:
     Visitor& operator=(const Visitor&); ///< No assignment operator
