+++ code/time_type_switch3.cpp
@@ -113,7 +113,7 @@ struct VisitorFor2 : ShapeVisitor
 
     const shape_kind<N>& first;
     const shape_kind<M>& second;
-    int                  result; 
+    size_t               result;
 
 private:
     VisitorFor2& operator=(const VisitorFor2&); ///< No assignment operator
@@ -134,7 +134,7 @@ struct VisitorFor : ShapeVisitor
 
     const shape_kind<N>& first;
     const Shape&         third;
-    int                  result; 
+    size_t               result;
 
 private:
     VisitorFor& operator=(const VisitorFor&); ///< No assignment operator
@@ -154,7 +154,7 @@ struct Visitor : ShapeVisitor
 
     const Shape& second;
     const Shape& third;
-    int          result;
+    size_t       result;
 
 private:
     Visitor& operator=(const Visitor&); ///< No assignment operator
